+++ src/amqp/src/Event/ConsumeEvent.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://doc.hyperf.io
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
+ */
+
+namespace Hyperf\Amqp\Event;
+
+use Hyperf\Amqp\Message\ConsumerMessageInterface;
+
+class ConsumeEvent
+{
+    /**
+     * @var ConsumerMessageInterface
+     */
+    protected $message;
+
+    public function __construct(ConsumerMessageInterface $message)
+    {
+        $this->message = $message;
+    }
+
+    public function getMessage(): ConsumerMessageInterface
+    {
+        return $this->message;
+    }
+}