+++ src/utils/src/Parallel.php
@@ -12,12 +12,29 @@ declare(strict_types=1);
 
 namespace Hyperf\Utils;
 
+use Swoole\Coroutine\Channel;
+
 class Parallel
 {
     /**
      * @var callable[]
      */
     private $callbacks = [];
+
+    /**
+     * @var null|Channel
+     */
+    private $concurrentChannel;
+
+    /**
+     * @param int $concurrent if $concurrent is equal to 0, that means unlimit
+     */
+    public function __construct(int $concurrent = 0)
+    {
+        if ($concurrent > 0) {
+            $this->concurrentChannel = new Channel($concurrent);
+        }
+    }
 
     public function add(callable $callable, $key = null)
     {
@@ -34,8 +51,10 @@ class Parallel
         $wg = new WaitGroup();
         $wg->add(count($this->callbacks));
         foreach ($this->callbacks as $key => $callback) {
+            $this->concurrentChannel && $this->concurrentChannel->push(true);
             Coroutine::create(function () use ($callback, $key, $wg, &$result) {
                 $result[$key] = call($callback);
+                $this->concurrentChannel && $this->concurrentChannel->pop();
                 $wg->done();
             });
         }