+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # v1.1.3
 
+## Added
+
+- [#745](https://github.com/hyperf-cloud/hyperf/pull/745) Added option `with-comments` for command `gen:model`.
+
 ## Fixed
 
 - [#741](https://github.com/hyperf-cloud/hyperf/pull/741) Fixed `db:seed` without filename.