+++ CHANGELOG.md
@@ -2,13 +2,13 @@
 
 ## Added
 
-- [#745](https://github.com/hyperf-cloud/hyperf/pull/745) Added option `with-comments` for command `gen:model`.
+- [#745](https://github.com/hyperf/hyperf/pull/745) Added option `with-comments` for command `gen:model`.
 - [#762](https://github.com/hyperf/hyperf/pull/762) Add concurrent for parallel.
 
 ## Fixed
 
-- [#741](https://github.com/hyperf-cloud/hyperf/pull/741) Fixed `db:seed` without filename.
-- [#748](https://github.com/hyperf-cloud/hyperf/pull/748) Fixed bug that `SymfonyNormalizer` not denormalize result of type `array`.
+- [#741](https://github.com/hyperf/hyperf/pull/741) Fixed `db:seed` without filename.
+- [#748](https://github.com/hyperf/hyperf/pull/748) Fixed bug that `SymfonyNormalizer` not denormalize result of type `array`.
 
 # v1.1.2 - 2019-10-17