+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 - [#748](https://github.com/hyperf/hyperf/pull/748) Fixed bug that `SymfonyNormalizer` not denormalize result of type `array`.
 - [#769](https://github.com/hyperf/hyperf/pull/769) Fixed invalid response exception throwed when result/error of jsonrpc response is null.
 
+# Changed
+
+- [#767](https://github.com/hyperf/hyperf/pull/767) Renamed property `running` to `listening` for `AbstractProcess`.
+
 # v1.1.2 - 2019-10-17
 
 ## Added