+++ CHANGELOG.md
@@ -1,11 +1,9 @@
 # v1.1.3
-## Added
-
-- [#747](https://github.com/hyperf/hyperf/pull/747) Added `AfterConsume`,`BeforeConsume`,`FailToConsume` events for AMQP consumer. 
 
 ## Added
 
 - [#745](https://github.com/hyperf/hyperf/pull/745) Added option `with-comments` for command `gen:model`.
+- [#747](https://github.com/hyperf/hyperf/pull/747) Added `AfterConsume`,`BeforeConsume`,`FailToConsume` events for AMQP consumer. 
 - [#762](https://github.com/hyperf/hyperf/pull/762) Add concurrent for parallel.
 
 ## Fixed